Move namespace viewsets to UI internal space
