+++ galaxy_api/api/urls.py
@@ -2,10 +2,10 @@
 
 from django.urls import include, path
 
-from .v3 import urls as urls_v3
+from .ui import urls as ui_urls
 
 
 app_name = 'api'
 urlpatterns = [
-    path('v3/', include(urls_v3, namespace='v3')),
+    path('v3/_ui/', include(ui_urls, namespace='ui')),
 ]
